+++ mastodon-s3/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Version bump for new base image
 * Update checkout for changed gemfile
+* Manually install json-canonicalization 0.3.1
 
 ---